Relocate helper CMs to tools.automation.release
